# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CampusRide API
  description: |
    CampusRide backend: REST API for inter-campus ride sharing.

    Response envelope (all endpoints):
    - success: `{ "success": true, "data": {...}, "meta": null }`
    - error: `{ "success": false, "error": { "code": "...", "message": "...", "fields": [...] } }`

    Common error codes: `VALIDATION_ERROR`, `NOT_FOUND`, `UNAUTHORIZED`, `FORBIDDEN`,
    `EMAIL_TAKEN`, `STUDENT_ID_TAKEN`, `AUTH_INVALID_CREDENTIALS`, `AUTH_TOKEN_INVALID`,
    `AUTH_REFRESH_REVOKED`, `ACCOUNT_SUSPENDED`, `DRIVER_NOT_VERIFIED`,
    `VEHICLE_NOT_VERIFIED`, `CAMPUS_NOT_FOUND`, `RIDE_NOT_FOUND`, `RIDE_FULL`,
    `RIDE_INVALID_STATE`, `BOOKING_NOT_FOUND`, `BOOKING_ALREADY_EXISTS`,
    `BOOKING_INVALID_STATE`, `INVALID_PIN`, `LOCATION_NOT_ALLOWED`, `RATE_LIMITED`,
    `INTERNAL_ERROR`.

    Authenticate with `Authorization: Bearer <accessToken>`. Tokens are issued by
    `POST /auth/login` / `POST /auth/register` and rotated by `POST /auth/refresh`.
  version: 0.1.0
  contact:
    name: CampusRide Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /rides:
    post:
      summary: Publish a ride (verified driver, verified active vehicle)
      operationId: createRide
      tags: [Rides]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRideRequest"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Error"
  /rides/search:
    get:
      summary: Search open rides between two campuses on a date, scored by matching
      operationId: searchRides
      tags: [Rides]
      parameters:
        - name: fromCampusId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: toCampusId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: true
          schema:
            type: string
            pattern: ^\d{4}-\d{2}-\d{2}$
        - name: seats
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
        - name: startTime
          in: query
          schema:
            type: string
            pattern: ^\d{2}:\d{2}$
        - name: endTime
          in: query
          schema:
            type: string
            pattern: ^\d{2}:\d{2}$
        - name: maxBudget
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /rides/{rideId}:
    get:
      summary: Get a ride detail (participants)
      operationId: getRide
      tags: [Rides]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    patch:
      summary: Update an open ride (driver only)
      operationId: updateRide
      tags: [Rides]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRideRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /rides/{rideId}/cancel:
    post:
      summary: Cancel an open/full ride (driver only)
      operationId: cancelRide
      tags: [Rides]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /rides/{rideId}/start:
    post:
      summary: Start a ride (driver only)
      operationId: startRide
      tags: [Rides]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                checkinAll:
                  type: boolean
                  description: Check in all accepted passengers immediately
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /rides/{rideId}/complete:
    post:
      summary: Complete a started ride (driver only)
      operationId: completeRide
      tags: [Rides]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /rides/{rideId}/location/latest:
    get:
      summary: Latest driver location for a ride
      operationId: latestLocation
      tags: [Rides]
      parameters:
        - $ref: "#/components/parameters/RideIdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RideIdParam:
      name: rideId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Success:
      description: Successful response with the standard envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiSuccess"
    Error:
      description: Failed response with the standard envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiSuccess:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          nullable: true
        meta:
          nullable: true
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message, fields]
          properties:
            code:
              type: string
              example: RIDE_FULL
            message:
              type: string
            fields:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string
              nullable: true
...
    CreateRideRequest:
      type: object
      required: [vehicleId, originCampusId, destinationCampusId, origin, destination, departureAt, availableSeats, pricePerSeat]
      properties:
        vehicleId:
          type: string
          format: uuid
        originCampusId:
          type: string
          format: uuid
        destinationCampusId:
          type: string
          format: uuid
        origin:
          type: string
          maxLength: 120
        destination:
          type: string
          maxLength: 120
        departureAt:
          type: string
          format: date-time
        expectedArrivalAt:
          type: string
          format: date-time
          nullable: true
        availableSeats:
          type: integer
          minimum: 1
          maximum: 20
        pricePerSeat:
          type: number
          minimum: 0
        pickupRadiusKm:
          type: number
          minimum: 0
          maximum: 50
          default: 2
        maxDetourMinutes:
          type: integer
          minimum: 0
          maximum: 240
          default: 30
        conditions:
          type: array
          items:
            type: string
        notes:
          type: string
          maxLength: 500
        recurring:
          type: boolean
        recurringDays:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
    UpdateRideRequest:
      type: object
      properties:
        departureAt:
          type: string
          format: date-time
        expectedArrivalAt:
          type: string
          format: date-time
          nullable: true
        pricePerSeat:
          type: number
          minimum: 0
        availableSeats:
          type: integer
          minimum: 1
          maximum: 20
          description: Cannot go below the number of booked seats.
        pickupRadiusKm:
          type: number
          minimum: 0
          maximum: 50
        maxDetourMinutes:
          type: integer
          minimum: 0
          maximum: 240
        conditions:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
...